requirements.txt
brotli==1.0.9
ConfigParser==5.2.0
cryptography==2.1.4
Cython==0.29.26
dl==0.1.0
docutils==0.18.1
Flask==2.0.2
Flask_Cors==3.0.10
HTMLParser==0.0.2
importlib_metadata==4.9.0
jnius==1.1.0
keyring==10.6.0
lockfile==0.12.2
lxml==4.7.1
numpy==1.21.4
ordereddict==1.1
protobuf==3.19.1
pyOpenSSL==21.0.0
python-dotenv==0.19.2
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
xmlrpclib==1.0.1
zipp==3.6.0
